updated documentation and refactored
