+++ docs/index.html
@@ -72,13 +72,13 @@ high school Sophmores, Jack and Evan.</p>
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:32 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
