+++ docs/lib_site.js.html
@@ -33,7 +33,7 @@
  * of the actaul site so that developers can see the change made
  * by those who are not developers. Built using 
  * {@link https://www.npmjs.com/package/nodegit|NodeGit}
- * 
+ * @author Evan Coulson &lt;ivansonofcoul@gmail.com>
  * @module lib/Site
  * 
  * 
@@ -49,7 +49,6 @@ const path = require('path');
 const fs = require('fs');
 const util = require('../util');
 const Logger = require('../util/Logger');
-const Crypto = require('crypto');
 
 const sitePath = path.resolve(__dirname, '../../site');
 const logger = Logger('Site.js', ['error']);
@@ -115,7 +114,7 @@ async function saveFile(data, filePath, message) {
 			//add to tree
 			logger.info('Adding...');
 			try {
-				filePath = localizePath(filePath);
+				filePath = util.localizePath(filePath);
 				logger.debug(filePath);
 				await index.addByPath(filePath.trim());
 				await index.write();
@@ -233,7 +232,7 @@ async function resolveConflicts(parent, repo, index) {
 		entry.parent = parent;
 		if (entry.isBlob()) {
 			const data = (await entry.getBlob()).toString();
-			const filePath = globalizePath(path.join(newTree.path(), entry.name()));
+			const filePath = util.globalizePath(path.join(newTree.path(), entry.name()));
 			fs.writeFileSync(filePath, data, 'utf-8');
 		}
 	});
@@ -364,43 +363,6 @@ async function createHeadReference(repo) {
 }
 
 /**
- * @description takes the global file path of a file within
- * the local site stored and resolves the string to pertain to the ./site
- * dir as the root of the path
- * @param {String} filePath path to file
- * @returns {String} returns localized string to the ./site root
- * @private
- */
-function localizePath(filePath) {
-	let subpaths = filePath.split('/');
-	for (let i = 0; i &lt; subpaths.length; i++) {
-		if (subpaths[i] == 'site') {
-			subpaths.splice(0, i + 1);
-			break;
-		}
-	}
-	filePath = '';
-	for (let i = 0; i &lt; subpaths.length; i++) {
-		if (i != 0) {
-			filePath += '/' + subpaths[i];
-		} else {
-			filePath += subpaths[i];
-		}
-	}
-	return filePath;
-}
-
-/**
- * @description returns the global file path
- * @param {String} filePath takes in the local site file path
- * @returns {String} returns the global file path of a local path
- * @private
- */
-function globalizePath(filePath) {
-	return path.resolve(__dirname, '../../site', filePath);
-}
-
-/**
  * @private
  * @description handles github credentials when publishing to repo
  * @param {Object} config app configuration object
@@ -413,42 +375,12 @@ async function handleCreds(config) {
 		logger.info('handling git hub credentials...');
 		const pass = config.git.password;
 		const user = config.git.username;
-		decrypt(config, user, function handleUsername(username) {
-			decrypt(config, pass, function hanldePassword(password) {
+		util.decrypt(config, user, function handleUsername(username) {
+			util.decrypt(config, pass, function hanldePassword(password) {
 				return resolve([username, password]);
 			});
 		});
 	});
-}
-
-/**
- * 
- * @param {Object} config app configuration object
- * @param {String} str string to decrypt
- * @param {Function} next callback when finished decrypting
- * @returns {String} returns the decrypted string to the
- * callback
- * @async
- * @private
- */
-function decrypt(config, str, next) {
-	logger.info('decrypting...');
-	const decipher = Crypto.createDecipher('aes192', config.secret);
-	let decrypted = '';
-
-	decipher.on('readable', function handleRead() {
-		const data = decipher.read();
-		if (data) {
-			decrypted += data.toString('utf8');
-		}
-	});
-
-	decipher.on('end', function handleClose() {
-		logger.info('finished decrypting...');
-		return next(decrypted);
-	});
-	decipher.write(str, 'hex');
-	decipher.end();
 }
 
 /**
@@ -481,13 +413,13 @@ module.exports = {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:32 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
