+++ docs/module-lib_Site.html
@@ -82,6 +82,13 @@ by those who are not developers. Built using
     
 
     
+    <dt class="tag-author">Author:</dt>
+    <dd class="tag-author">
+        <ul>
+            <li><a href="mailto:ivansonofcoul@gmail.com">Evan Coulson</a></li>
+        </ul>
+    </dd>
+    
 
     
 
@@ -273,7 +280,7 @@ by those who are not developers. Built using
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line431">line 431</a>
+        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line363">line 363</a>
     </li></ul></dd>
     
 
@@ -431,7 +438,7 @@ with any recent changes made by the development team
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line37">line 37</a>
+        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line36">line 36</a>
     </li></ul></dd>
     
 
@@ -547,7 +554,7 @@ publish remote and branch of the repo
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line115">line 115</a>
+        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line114">line 114</a>
     </li></ul></dd>
     
 
@@ -704,7 +711,7 @@ publish remote and branch of the repo
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line141">line 141</a>
+        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line140">line 140</a>
     </li></ul></dd>
     
 
@@ -908,7 +915,7 @@ file and adds and commits file to recent changes
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line73">line 73</a>
+        <a href="lib_site.js.html">lib/site.js</a>, <a href="lib_site.js.html#line72">line 72</a>
     </li></ul></dd>
     
 
@@ -959,13 +966,13 @@ file and adds and commits file to recent changes
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:33 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
