+++ docs/module-util_Logger.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="utf-8">
-    <title>JSDoc: Class: module:util/Logger</title>
+    <title>JSDoc: Module: util/Logger</title>
 
     <script src="scripts/prettify/prettify.js"> </script>
     <script src="scripts/prettify/lang-css.js"> </script>
@@ -17,7 +17,7 @@
 
 <div id="main">
 
-    <h1 class="page-title">Class: module:util/Logger</h1>
+    <h1 class="page-title">Module: util/Logger</h1>
 
     
 
@@ -28,7 +28,10 @@
 
 <header>
     
-        <h2><span class="attribs"><span class="type-signature"></span></span>module:util/Logger<span class="signature">(tag, debugLevels, testingEnv)</span><span class="type-signature"> &rarr; {Logger:Object}</span></h2>
+        
+            
+        
+            
         
     
 </header>
@@ -37,11 +40,18 @@
     <div class="container-overview">
     
         
+            <div class="description">The logger module is built on top of the Winston logging module 
+<a href="https://www.npmjs.com/package/winston">Winston</a>.
+This module allows for loggers with different tags and debuging levels to be created with ease.</div>
+        
+
+        
+            
 
     
 
     
-    <h4 class="name" id="module:util/Logger"><span class="type-signature"></span>new module:util/Logger<span class="signature">(tag, debugLevels, testingEnv)</span><span class="type-signature"> &rarr; {Logger:Object}</span></h4>
+    <h4 class="name" id="module:util/Logger"><span class="type-signature"></span>(require("util/Logger"))<span class="signature">(tag, debugLevels, testingEnv)</span><span class="type-signature"> &rarr; {Logger:Object}</span></h4>
     
 
     
@@ -244,11 +254,99 @@ each transport
 
 
 
+        
+            
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+<dl class="details">
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-author">Author:</dt>
+    <dd class="tag-author">
+        <ul>
+            <li><a href="mailto:ivansonofcoul@gmail.com">Evan Coulson</a></li>
+        </ul>
+    </dd>
+    
+
+    
+
+    
+
+    
+
+    
+    <dt class="tag-source">Source:</dt>
+    <dd class="tag-source"><ul class="dummy"><li>
+        <a href="util_Logger.js.html">util/Logger.js</a>, <a href="util_Logger.js.html#line1">line 1</a>
+    </li></ul></dd>
+    
+
+    
+
+    
+
+    
+</dl>
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        
     
     </div>
 
     
 
+    
+        <h3 class="subsection-title">Requires</h3>
+
+        <ul>
+            <li>module:Winston</li>
+        </ul>
     
 
     
@@ -276,13 +374,13 @@ each transport
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:33 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
