+++ docs/module-util_MYSQL.html
@@ -83,6 +83,13 @@ database. This class requires the <a href="module-util_Logger.html">Logger</a>,
     
 
     
+    <dt class="tag-author">Author:</dt>
+    <dd class="tag-author">
+        <ul>
+            <li><a href="mailto:ivansonofcoul@gmail.com">Evan Coulson</a></li>
+        </ul>
+    </dd>
+    
 
     
 
@@ -229,7 +236,7 @@ database to connect to within MYSQL.</td>
     
     <dt class="tag-source">Source:</dt>
     <dd class="tag-source"><ul class="dummy"><li>
-        <a href="util_mysql.js.html">util/mysql.js</a>, <a href="util_mysql.js.html#line37">line 37</a>
+        <a href="util_mysql.js.html">util/mysql.js</a>, <a href="util_mysql.js.html#line38">line 38</a>
     </li></ul></dd>
     
 
@@ -326,13 +333,13 @@ Sequelize Schema types.
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:33 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
