+++ docs/util_Logger.js.html
@@ -30,7 +30,8 @@
  * @description The logger module is built on top of the Winston logging module 
  * {@link https://www.npmjs.com/package/winston|Winston}.
  * This module allows for loggers with different tags and debuging levels to be created with ease.
- *
+ * @author Evan Coulson &lt;ivansonofcoul@gmail.com>
+ * 
  * @module util/Logger
  * 
  * @requires Winston
@@ -62,7 +63,6 @@ const logFormat = printf(function handleFormat(info) {
  * running in a testing env or not
  * @returns {Logger:Object} Returns a logger that logs all given values to
  * each transport
- * @constructor
  */
 module.exports = function Logger(tag, debugLevels, testingEnv) {
 	let fileArray = debugLevels.map((x) => {
@@ -124,13 +124,13 @@ module.exports = function Logger(tag, debugLevels, testingEnv) {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:32 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
