+++ docs/util_index.js.html
@@ -26,9 +26,15 @@
     
     <section>
         <article>
-            <pre class="prettyprint source linenums"><code>const fs = require('fs');
+            <pre class="prettyprint source linenums"><code>/**
+ * @description Implements util functions for all modules in the library
+ * @author Evan Coulson &lt;ivansonofcoul@gmail.com>
+ */
+
+const fs = require('fs');
 const path = require('path');
 const Logger = require('../util/Logger');
+const Crypto = require('crypto');
 
 const logger = Logger('lib.js', ['error', 'debug']);
 
@@ -52,8 +58,79 @@ function readConfig() {
 	});
 }
 
+
+/**
+ * @description takes the global file path of a file within
+ * the local site stored and resolves the string to pertain to the ./site
+ * dir as the root of the path
+ * @param {String} filePath path to file
+ * @returns {String} returns localized string to the ./site root
+ * @private
+ */
+function localizePath(filePath) {
+	let subpaths = filePath.split('/');
+	for (let i = 0; i &lt; subpaths.length; i++) {
+		if (subpaths[i] == 'site') {
+			subpaths.splice(0, i + 1);
+			break;
+		}
+	}
+	filePath = '';
+	for (let i = 0; i &lt; subpaths.length; i++) {
+		if (i != 0) {
+			filePath += '/' + subpaths[i];
+		} else {
+			filePath += subpaths[i];
+		}
+	}
+	return filePath;
+}
+
+/**
+ * @description returns the global file path
+ * @param {String} filePath takes in the local site file path
+ * @returns {String} returns the global file path of a local path
+ * @private
+ */
+function globalizePath(filePath) {
+	return path.resolve(__dirname, '../../site', filePath);
+}
+
+/**
+ * 
+ * @param {Object} config app configuration object
+ * @param {String} str string to decrypt
+ * @param {Function} next callback when finished decrypting
+ * @returns {String} returns the decrypted string to the
+ * callback
+ * @async
+ * @private
+ */
+function decrypt(config, str, next) {
+	logger.info('decrypting...');
+	const decipher = Crypto.createDecipher('aes192', config.secret);
+	let decrypted = '';
+
+	decipher.on('readable', function handleRead() {
+		const data = decipher.read();
+		if (data) {
+			decrypted += data.toString('utf8');
+		}
+	});
+
+	decipher.on('end', function handleClose() {
+		logger.info('finished decrypting...');
+		return next(decrypted);
+	});
+	decipher.write(str, 'hex');
+	decipher.end();
+}
+
 module.exports = {
 	readConfig: readConfig,
+	decrypt: decrypt,
+	globalizePath: globalizePath,
+	localizePath: localizePath,
 };</code></pre>
         </article>
     </section>
@@ -64,13 +141,13 @@ module.exports = {
 </div>
 
 <nav>
-    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Classes</h3><ul><li><a href="module-util_Logger.html">util/Logger</a></li></ul><h3>Global</h3><ul><li><a href="global.html#readConfig">readConfig</a></li></ul>
+    <h2><a href="index.html">Home</a></h2><h3>Modules</h3><ul><li><a href="module-lib_Site.html">lib/Site</a></li><li><a href="module-util_Logger.html">util/Logger</a></li><li><a href="module-util_MYSQL.html">util/MYSQL</a></li></ul><h3>Global</h3><ul><li><a href="global.html#fs">fs</a></li><li><a href="global.html#readConfig">readConfig</a></li></ul>
 </nav>
 
 <br class="clear">
 
 <footer>
-    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:44:55 GMT-0800 (PST)
+    Documentation generated by <a href="https://github.com/jsdoc3/jsdoc">JSDoc 3.5.5</a> on Sun Nov 26 2017 17:51:32 GMT-0800 (PST)
 </footer>
 
 <script> prettyPrint(); </script>
